Move oauth signals into oauth.py
