+++ aleph/oauth.py
@@ -1,6 +1,9 @@
 from flask_oauthlib.client import OAuth
 from flask import session
 
+from aleph import signals
+from aleph.model import Role
+
 oauth = OAuth()
 
 
@@ -32,3 +35,31 @@ def configure_oauth(app):
     setup_providers(app)
     oauth.init_app(app)
     return oauth
+
+
+@signals.handle_oauth_session.connect
+def handle_google_oauth(sender, provider=None, session=None):
+    # If you wish to use another OAuth provider with your installation of
+    # aleph, you can create a Python extension package and include a
+    # custom oauth handler like this, which will create roles and state
+    # for your session.
+    if 'googleapis.com' not in provider.base_url:
+        return
+    me = provider.get('userinfo')
+    user_id = 'google:%s' % me.data.get('id')
+    role = Role.load_or_create(user_id, Role.USER, me.data.get('name'),
+                               email=me.data.get('email'))
+    session['roles'].append(role.id)
+    session['user'] = role.id
+
+
+@signals.handle_oauth_session.connect
+def handle_facebook_oauth(sender, provider=None, session=None):
+    if 'facebook.com' not in provider.base_url:
+        return
+    me = provider.get('me?fields=id,name,email')
+    user_id = 'facebook:%s' % me.data.get('id')
+    role = Role.load_or_create(user_id, Role.USER, me.data.get('name'),
+                               email=me.data.get('email'))
+    session['roles'].append(role.id)
+    session['user'] = role.id
