+++ aleph/views/sessions_api.py
@@ -97,34 +97,6 @@ def logout():
     return redirect('/')
 
 
-@signals.handle_oauth_session.connect
-def handle_google_oauth(sender, provider=None, session=None):
-    # If you wish to use another OAuth provider with your installation of
-    # aleph, you can create a Python extension package and include a
-    # custom oauth handler like this, which will create roles and state
-    # for your session.
-    if 'googleapis.com' not in provider.base_url:
-        return
-    me = provider.get('userinfo')
-    user_id = 'google:%s' % me.data.get('id')
-    role = Role.load_or_create(user_id, Role.USER, me.data.get('name'),
-                               email=me.data.get('email'))
-    session['roles'].append(role.id)
-    session['user'] = role.id
-
-
-@signals.handle_oauth_session.connect
-def handle_facebook_oauth(sender, provider=None, session=None):
-    if 'facebook.com' not in provider.base_url:
-        return
-    me = provider.get('me?fields=id,name,email')
-    user_id = 'facebook:%s' % me.data.get('id')
-    role = Role.load_or_create(user_id, Role.USER, me.data.get('name'),
-                               email=me.data.get('email'))
-    session['roles'].append(role.id)
-    session['user'] = role.id
-
-
 @blueprint.route('/api/1/sessions/callback/<string:provider>')
 def callback(provider):
     oauth_provider = oauth.remote_apps.get(provider)
